+++ backend/core/src/seeder/seeds/listings/listing-triton-seed.ts
@@ -90,7 +90,7 @@ const tritonListing: ListingSeedType = {
   name: "Test: Triton",
   postmarkedApplicationsReceivedByDate: null,
   programRules: null,
-  rentalAssistance: "",
+  rentalAssistance: "Rental assistance",
   rentalHistory: "No evictions",
   requiredDocuments:
     "Due at interview - Paystubs, 3 months’ bank statements, recent tax returns or non-tax affidavit, recent retirement statement, application to lease, application qualifying criteria, social security card, state or nation ID. For self-employed, copy of IRS Tax Return including schedule C and current or most recent clients. Unemployment if applicable. Child support/Alimony; current notice from DA office, a court order or a letter from the provider with copies of last two checks. Any other income etc",