+++ sites/public/pages/create-account.tsx
@@ -62,6 +62,9 @@ export default () => {
       const { status, data } = err.response || {}
       if (status === 400) {
         setRequestError(`${t(`authentication.createAccount.errors.${data.message}`)}`)
+      } else if (status === 409) {
+        console.error(err)
+        setRequestError(`${t("authentication.createAccount.errors.emailInUse")}`)
       } else {
         console.error(err)
         setRequestError(`${t("authentication.createAccount.errors.generic")}`)
@@ -214,14 +217,7 @@ export default () => {
             />
 
             <div className="text-center mt-10">
-              <Button
-                styleType={AppearanceStyleType.primary}
-                onClick={() => {
-                  console.info("button has been clicked!")
-                }}
-              >
-                {t("account.createAccount")}
-              </Button>
+              <Button styleType={AppearanceStyleType.primary}>{t("account.createAccount")}</Button>
             </div>
           </div>
         </Form>